Adding files used to compare speech vs text
This was built using Microsoft Copilot:
Prompt 1: I have a use case that involves creating an HTML file that has a few phrases that I want to practice saying in a microphone. The HTML file needs to have the capability to do STT (speech-to-text) where the results of the speech would be displayed. With the use of HTML and JavaScript code, can you help me by providing an output that would create my use case?
Follow up Prompt (Prompt 2): Thanks, is it possible that you can tweak the following code to do the following:
- Make the 3 phrases that were given earlier as radio button options
- Add a button labeled 'Get Score' that outputs a score (0 to 100) on how accurate the