Adding HTML application to directory
Prompt: Can you build me an HTML application that plays a video in the top-center of the page (the video will be from a URL link that I'll provide later) and has a textbox and a submit button that will alert the user that their answer is correct if they put in one of the following inputs: "knight", "Knight", "The Knight", "the knight". Otherwise, the user will get an alert saying their answer is incorrect. Can you help build this application by using HTML, CSS, and JavaScript code?

Output: Below is a complete HTML, CSS, and JavaScript application that meets your criteria. Just replace "YOUR_VIDEO_URL_HERE.mp4" with your actual video URL when you're ready: This code creates an application where a video plays at the top-center of the page, and below it you have a textbox and a submit button. When the user enters one of the designated answers ("knight", "Knight", "The Knight", or "the knight"), they'll receive a "Correct!" alert; any other input will prompt an "Incorrect." alert.
Would you like any further modifications or additional features to enhance the interactivity or design of this application?